The New Line
You are familiar with producing output using CPP. In this task, you'll be required to write three messages in three separate lines. Write Geeks for Geeks. But wait a minute!! You need to write each word of it in a separate line.
In the function printInNewLine(), output each word of Geeks for Geeks in a separate line.
Hint: endl might be useful here!

Input:
No input

Output:
Geeks
for
Geeks

Explanation:
First word of the statement is in first line, next word is in next line, and last is in last line.